New Hover Fix
Instead of using `pointer-events: none` which prevents users from using right-click -> Open in New Tab, I came up with a new fix which uses a masking approach. Over each rectangle in the grid is a pseudo-element which sits at z-index 2. When you hover on an image, it's set to z-index 1, while the pseudo-element that would otherwise overlay it is set to z-index -1.

This results in a sort of mask, where the only place the pointer is directly over the expanded image is when it's over its original grid-space, and moving to an adjacent grid space will put the pointer over that space's pseudo-element mask, causing the corresponding image to be hovered instead.

This fix allowed me to remove the `pointer-events: none` and return the title attributes to the images, where they belong.