Adding Links and Fixing Rapid Flipping Bug
I added links to the character picker, the GitHub repo, and my Twitter. The link menu is a flex row itself, so it will collapse and flatten as needed. Not ideal, but good enough.

There was a bug previously where you could cause 2 images to rapidly alternate expanding and shrinking by placing the cursor on the edge between them. I fixed this by making the masks very slightly smaller than the images.

This introduces a potential bug where you can avoid changing which image you're hovering over by weaving between the masks, but this requires inhuman precision for a hardly-noticeable effect. I much prefer this bug to the much more obnoxious, easier-to-trigger bug that it replaces.