Improving support for touch devices
This commit changes the behavior of the app when using a touch screen. Now, the zoomed image will follow your finger as you drag it across the grid. Also, double-tap to zoom has been disabled, which allows users to repeatedly tap buttons (like the Random Stage button) without the gesture causing the device to zoom erratically.

This implementation may change, as more testing should be done to assure it is as accessible as possible on all devices.