Adding pointer-events: none to image overlay, making right click not change selection type
This allows you to still get the title text of or right click on the stage images, even when they have a selection overlay on them (the check mark, x mark, and question mark. The second change makes it easier to right click to perform secondary actions on images, without messing up the selection types.